Complete the registering using blueprint for the other frinedns
