+++ templates/about.html
@@ -31,16 +31,26 @@
 </head>
 
 <body>
-    <section class="cont">
-        {% for user in people %}
-        <div>
-            <img src="{{ user.pic }}" alt="">
-            <h1>Hello, {{ user.name }}</h1>
-        </div>
-        {% endfor %}
-    </section>
-
-
+    <nav>
+        <ul>
+          <li><a href="/profile"> Profile </a></li>
+          <li><a href="/about"> About </a></li>
+          <li><a href="/"> Home </a></li>
+        </ul>
+      </nav>
+      
+      <!-- mustache -->
+      {% for user in users %}
+      <div class="user-container">
+        <img class="user-pic" src="{{user.pic}}" alt="Profile picture" />
+        <p class="user-msg">
+          Hello,
+          <span class="user-name"
+            >{{user.name}} {{ "🚀" if user.pro else "🎊🎊" }}
+          </span>
+        </p>
+      </div>
+      {% endfor %}
 </body>
 
 </html>
